feat(fluxWizard): introduce timeRangeWindowPeriod
