+++ ui/src/shared/components/TimeMachine/fluxQueryBuilder/util/timeRangeWindowPeriod.ts
@@ -1,0 +1,18 @@
+import moment from 'moment'
+import {fluxPeriodFromRangeSeconds} from 'src/tempVars/utils/replace'
+import {TimeRange} from 'src/types'
+
+function timeRangeWindowPeriod(timeRange: TimeRange | undefined) {
+  if (timeRange) {
+    if (timeRange.seconds) {
+      return fluxPeriodFromRangeSeconds(timeRange.seconds)
+    }
+    // calculate from upper / lower
+    const seconds = Math.round(
+      moment(timeRange.upper).diff(moment(timeRange.lower)) / 1000
+    )
+    return fluxPeriodFromRangeSeconds(seconds)
+  }
+}
+
+export default timeRangeWindowPeriod
